fix: move LaunchPlanLink to component/Launchplan
